// src/types/splashtail/silverpelt.ts
export type CommandExtendedDataMap = Record<string, CommandExtendedData>;
export interface CanonicalModule {
	id: string;
	name: string;
	description: string;
	toggleable: boolean;
	commands_toggleable: boolean;
	web_hidden: boolean;
	virtual_module: boolean;
	is_default_enabled: boolean;
	commands: CanonicalCommand[];
	s3_paths: string[];
	config_options: CanonicalConfigOption[];
}
export interface CanonicalCommand {
	command: CanonicalCommandData;
	extended_data: CommandExtendedDataMap;
}
export interface CanonicalCommandArgument {
	name: string;
	description?: string;
	required: boolean;
	choices: string[];
}
export interface CanonicalCommandData {
	name: string;
	qualified_name: string;
	description?: string;
	nsfw: boolean;
	subcommands: CanonicalCommandData[];
	subcommand_required: boolean;
	arguments: CanonicalCommandArgument[];
}
// ...
export interface CanonicalColumn {
	id: string;
	name: string;
	description: string;
	column_type: CanonicalColumnType;
	nullable: boolean;
	suggestions: CanonicalColumnSuggestion;
	secret: boolean;
	ignored_for: CanonicalOperationType[];
}
export type CanonicalOperationType = string;
export const View: CanonicalOperationType = 'View';
export const Create: CanonicalOperationType = 'Create';
export const Update: CanonicalOperationType = 'Update';
export const Delete: CanonicalOperationType = 'Delete';
export interface CanonicalConfigOption {
	id: string;
	name: string;
	description: string;
	primary_key: string;
	title_template: string;
	columns: CanonicalColumn[];
	operations: CanonicalOperationType[];
}

//////////
// source: db__interface.go

export type DbConn = any;

//////////
// source: mod.go
/*
From silverpelt/mod.rs
*/

/**
 * PermissionCheck represents the permissions needed to run a command.
 */
export interface PermissionCheck {
	kittycat_perms: string[]; // The kittycat permissions needed to run the command
	native_perms: string /* bigint */[]; // The native permissions needed to run the command (converted from serenity::all::Permissions)
	inner_and: boolean; // Whether or not the perms are ANDed (all needed) or OR'd (at least one)
}
/**
 * CommandExtendedData represents the default permissions needed to run a command.
 */
export interface CommandExtendedData {
	default_perms: PermissionCheck; // The default permissions needed to run this command
	is_default_enabled: boolean; // Whether or not the command is enabled by default
	web_hidden: boolean; // Whether or not the command is hidden from the web interface
	virtual_command: boolean; // Whether or not the command is a virtual command or not
}